backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
httpx==0.25.2
aiohttp==3.9.1
apscheduler==3.10.4
python-telegram-bot==20.7
python-multipart==0.0.6
pydantic-settings==2.1.0
python-dateutil==2.9.0
slowapi==0.1.9